Experiment with data.table syntax
In an attempt to rid my life of the family of apply functions, I switched to the equally--if not more--horrendous data.table syntax. Mainly as a way to speed up the grep family of functions in enc_check.R. Need to rewrite ASCIIreplace.R to take into account the slightly different structure of the returned object from the data.table-based  enc_check function.