Patch control_replace.R. Begin new function.
control_replace.R used a fixed variable, map_matches, which stems from the /tests/Munge.R script. Replaced with proper script inputs.
Currently writing function to return the maximum number of
sequences in a column, so that users don't have to guess at the
number of columns to use in their repalcement matrix.